+++ src/main/java/ru/chertenok/webapps/webstore/bd/daoimpl_bd/AbstractDAOimpl.java
@@ -1,0 +1,15 @@
+package ru.chertenok.webapps.webstore.bd.daoimpl_bd;
+
+import org.jetbrains.annotations.NotNull;
+
+import javax.persistence.EntityManager;
+import javax.persistence.PersistenceContext;
+
+public abstract class AbstractDAOimpl {
+    @NotNull
+    public static final String PERSISTENCE_UNIT = "MYSQLPersistence";
+
+    @PersistenceContext(unitName = PERSISTENCE_UNIT)
+    protected EntityManager em;
+
+}
