--- src/main/webapp/WEB-INF/web.xml
@@ -1,57 +1,0 @@
-<web-app xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xmlns="http://xmlns.jcp.org/xml/ns/javaee"
-         xsi:schemaLocation="http://xmlns.jcp.org/xml/ns/javaee
-         http://xmlns.jcp.org/xml/ns/javaee/web-app_3_1.xsd"
-         version="3.1">
-
-    <display-name>Web Store App</display-name>
-
-    <welcome-file-list>
-        <welcome-file>index</welcome-file>
-    </welcome-file-list>
-
-
-    <context-param>
-        <param-name>PATH_JSP</param-name>
-        <param-value>/WEB-INF/JSP/</param-value>
-    </context-param>
-    <context-param>
-        <param-name>SHABLON</param-name>
-        <param-value>shablon.jsp</param-value>
-    </context-param>
-    <context-param>
-        <param-name>javax.faces.PROJECT_STAGE</param-name>
-        <param-value>Development​</param-value>
-    </context-param>
-    <context-param>
-        <param-name>primefaces.FONT_AWESOME</param-name>
-        <param-value>true</param-value>
-    </context-param>
-
-    <listener>
-        <listener-class>com.sun.faces.config.ConfigureListener</listener-class>
-    </listener>
-
-    <servlet>
-        <servlet-name>faces</servlet-name>
-        <servlet-class>javax.faces.webapp.FacesServlet</servlet-class>
-        <load-on-startup>1</load-on-startup>
-    </servlet>
-
-
-    <servlet-mapping>
-        <servlet-name>faces</servlet-name>
-        <url-pattern>*.xhtml</url-pattern>
-    </servlet-mapping>
-
-    <error-page>
-        <error-code>404</error-code>
-        <location>/WEB-INF/JSP/error404.jsp</location>
-    </error-page>
-    <error-page>
-        <error-code>403</error-code>
-        <location>/WEB-INF/JSP/error403.jsp</location>
-    </error-page>
-
-
-</web-app>
