+++ src/main/webapp/admin/item-view.xhtml
@@ -58,8 +58,9 @@
                     <h:outputText value="Вариант"/>
                 </f:facet>
 
-                <p:dataList id="varList" value="#{item.listVariants}" var="itemVar" type="unordered"
-                            rendered="#{itemViewController.itemDAO.items.size()>0}"
+                <p:dataList id="varList" value="#{itemViewController.getVariantByItem(item)}" var="itemVar"
+                            type="unordered"
+
                             itemType="none" paginator="true" paginatorAlwaysVisible="false">
                     <p:link outcome="/admin">
                         <i class="fa fa-edit" style="font-size: 0.8em; margin: 5px;"/>
@@ -80,11 +81,6 @@
             </p:column>
 
         </p:dataTable>
-
-            <p:blockUI block="dataTable" trigger="dataTable">
-                LOADING<br/>
-                <p:graphicImage name="demo/images/ajaxloadingbar.gif"/>
-            </p:blockUI>
 
 
     </h:form>
